build: rework setup of Rust toolchains
The setup of all tools is now done via a Gradle task. The tools
are installed locally in the 'build' directory of the project
that requires them. They are thus isolated from any state of
Rust tooling already installed on a (local) machine.

The 'installRustToolchain' task is cached. Which means that the
complete toolchain is put into the Gradle remote cache. Thus, once
the installation ran successfully on a CI agent of a given
operating system, the ready-to-use toolchain will be downloaded from
the cache if the build runs on a machine with that operating system.

For stability and reproducibility, for both local and CI builds, the
versions of Rust and the additional toolchain components (zig, xwin, ...)
are now pinned and require explicit updating when desired.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>